Database Refinements
Comments, helper folder + small improvements.
