+++ dbDefinition.js
@@ -1,3 +1,5 @@
+// Defines database. Could be changed to different database here if we want to use something other than SQLite.
+
 const Sequelize = require('sequelize');
 
 const sequelize = new Sequelize('database', 'username', 'password', {
