+++ dbObjects.js
@@ -1,3 +1,5 @@
+// Defines helper methods within database models.
+// For example, getting a user could call some method user.foo() if defined here.
 const { sequelize, dataTypes } = require("./dbDefinition")
 
 const Users = require('./models/Users.js')(sequelize, dataTypes);
